# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
  handlers.AccessTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  handlers.ActivateRequest:
    properties:
      confirm_password:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - confirm_password
    - password
    - token
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: jdoe@ankabanka.com
        type: string
      password:
        example: secret
        type: string
    type: object
  handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handlers.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: REST API gateway for EXBanka microservices.
  title: EXBanka API Gateway
  version: "1.0"
paths:
...
  /auth/activate:
    post:
      consumes:
      - application/json
      description: Activate a new employee account using an activation token and set
        a password.
      parameters:
      - description: Activation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Activate account
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset email to the given address.
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Set a new password using a password reset token.
      parameters:
      - description: Reset payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - auth
  /client/activate:
    post:
      consumes:
      - application/json
      description: Activate a new client account using an activation token and set
        a password.
      parameters:
      - description: Activation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Activate client account
      tags:
      - auth
  /client/login:
    post:
      consumes:
      - application/json
      description: Authenticate a client with email and password, receive JWT tokens.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Client login
      tags:
      - auth
  /client/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid client refresh token for a new access token.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh client access token
      tags:
      - auth
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, receive JWT tokens.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"